refactor(respecDocWriter): move write out, use in CLI directly
